+++ src/libs/satellite/src/usage/mod.rs
@@ -3,5 +3,4 @@ mod impls;
 mod state;
 pub mod store;
 pub mod types;
-pub mod user_usage;
 mod utils;
